#474: Moved the AbilityCopyEffectsParser to the model folder
 The file was also renamed, but the parser was not touched.
